license_check.py: Do not report 'unknown-license-reference' as failures
This commit updates the license check script to ignore the reported
license type of `unknown-license-reference` because the Scancode often
reports this type when there exists a line containing the word
"license" in a source file.

Note that `unknown-license-reference` is different from finding an
unknown license type (`unknown-spdx) or entirely missing a license
header (`licenses` is empty).

For more details, refer to the issue #14.

Signed-off-by: Stephanos Ioannidis <root@stephanos.io>