entrypoint.sh: exit early if git fails
Change the entrypoint script to configure the shell to exit on the first
command error and change the git call to set a local intermediate
variable, so that any failure in the git command makes the script exit
early with an error code.

This way if there's some problem with the checkout the script should
not fail silently.

Signed-off-by: Fabio Baltieri <fabiobaltieri@google.com>